requirements.txt
agora-token-builder==1.0.0
Cython==0.29.19
dj-database-url==1.0.0
Django==3.2.16
django-allauth==0.51.0
django-autoslug==1.9.8
django-browser-reload==1.6.0
django-compression-middleware==0.4.1
django-cors-headers==3.13.0 
django-filter==22.1
django-froala-editor==4.0.13
django-js-asset==2.0.0
django-storages==1.13.1
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.2.0
postgres==4.0
psycopg2-binary~=2.9.3
PyJWT==2.4.0
whitenoise==6.2.0
crispy-bootstrap4
django-crispy-forms
sqlparse==0.2.4
face_recognition==1.2.3